Use json.RawMessage for UpstreamMessage's Data field
We then parse TextMessages from there.

The big problem right now is that we're hitting FCM's limit on how big
messages can be. We're going to have to send the parts of the message
directly, store them somehow, and then send an id upstream along with
them. A bit annoying, but it's how it's going to have to be done...